Routing
Routing : Routing means navigating between pages. In angular the pages that we are referring to will be in the form of components.

In case of routing we have to create application using below command as
ng new Demo --routing

Due to this option app-routing.module.ts file gets created. In this file there is one array named as routes which contains information about routes.

By using wildcards we can provide the option if targeted URL is not valid.