Custom Directives
In angular like Inbuilt directives we can design our own custom directive directive while designing html form.

Custom directives are used in Angular to extend the functionality of HTML.

To create custom directive we have to use below command as
ng generate directive directive_name

Due to this command one .ts file gets created. In that file there is one class which gets exported and we can write our logic in that class which gets executed when we use the custom directive.