NgIf
Angular Structural Directives are directives which change the structure of the DOM by adding or removing elements.
There are three built in structural directives, NgIf, NgFor and NgSwitch.

These directives work by using the HTML5 <ng-template> tag. This is a new tag in HTML which is specifically designed to hold template code.